backend/requirements.txt
transformers
torch
sentencepiece
accelerate
pycryptodome
requests
uvicorn[standard]
passlib
python-jose
